+++ app/src/main/java/com/fieldbook/tracker/preferences/GeneralPreferencesFragment.java
@@ -29,7 +29,6 @@ public class GeneralPreferencesFragment extends PreferenceFragmentCompat impleme
     PreferenceManager prefMgr;
     Context context;
     private Preference defaultStorageLocation;
-    private KeyUtil mKeys;
 
     @Override
     public void onCreatePreferences(Bundle savedInstanceState, String rootKey) {
@@ -109,8 +108,6 @@ public class GeneralPreferencesFragment extends PreferenceFragmentCompat impleme
         // Occurs before the on create function. We get the context this way.
         GeneralPreferencesFragment.this.context = context;
 
-        mKeys = new KeyUtil(context);
-
     }
 
     @Override
@@ -127,54 +124,6 @@ public class GeneralPreferencesFragment extends PreferenceFragmentCompat impleme
         }
 
         return true;
-    }
-
-    private void setupTtsPreference() {
-
-        CheckBoxPreference ttsEnabled = findPreference(GeneralKeys.TTS_LANGUAGE_ENABLED);
-        if (ttsEnabled != null) {
-            Preference ttsLanguage = findPreference(GeneralKeys.TTS_LANGUAGE);
-            if (ttsLanguage != null) {
-                ttsLanguage.setVisible(ttsEnabled.isChecked());
-
-                ttsLanguage.setOnPreferenceClickListener((v) -> {
-                    startActivity(new Intent(getContext(), LocaleChoiceActivity.class));
-                    return true;
-                });
-
-                ttsLanguage.setOnPreferenceChangeListener((p, v) -> {
-                    int code = (int) v;
-                    updateTtsSummary(code);
-                    return true;
-                });
-            }
-
-            ttsEnabled.setOnPreferenceChangeListener((preference, newValue) -> {
-
-                if (ttsLanguage != null) {
-                    ttsLanguage.setVisible((Boolean) newValue);
-                }
-                return true;
-            });
-        }
-    }
-
-    private void updateTtsSummary(int code) {
-
-        Preference ttsLanguage = findPreference(GeneralKeys.TTS_LANGUAGE);
-
-        if (ttsLanguage != null) {
-
-            Set<Locale> locales = TextToSpeechHelper.Companion.getAvailableLocales();
-            for (Locale l : locales) {
-                if (l.hashCode() == code) {
-
-                    ttsLanguage.setSummary(l.getDisplayName());
-
-                    break;
-                }
-            }
-        }
     }
 
     @Override
@@ -194,10 +143,5 @@ public class GeneralPreferencesFragment extends PreferenceFragmentCompat impleme
 
             }
         }
-
-        setupTtsPreference();
-
-        int code = PreferenceManager.getDefaultSharedPreferences(context).getInt(mKeys.getArgTtsLocale(), -1);
-        updateTtsSummary(code);
     }
 }
