+++ app/src/main/res/values/strings.xml
@@ -735,5 +735,6 @@
     <string name="pref_profile_crashlytics_user_id_enabled_summary">This optional setting will log an anonymous, unique ID that will help the development team track and solve crashes.</string>
     <string name="pref_profile_crashlytics_refresh_title">Refresh ID</string>
     <string name="pref_profile_crashlytics_refresh_summary">This will assign a new unique ID.</string>
+    <string name="preferences_category_audio_tts_title">Text to Speech</string>
 
 </resources>
