+++ app/src/main/res/xml/preferences_general.xml
@@ -67,18 +67,4 @@
         android:summary="@string/preferences_general_feature_barcode_text_description"
         android:title="@string/preferences_general_feature_barcode_text" />
 
-    <CheckBoxPreference
-        android:defaultValue="false"
-        android:icon="@drawable/ic_tts"
-        android:key="TTS_LANGUAGE_ENABLED"
-        android:summary="@string/preferences_general_tts_description"
-        android:title="@string/preferences_general_tts_title" />
-
-    <Preference
-        app:isPreferenceVisible="false"
-        android:icon="@drawable/ic_pref_general_root_directory"
-        android:key="TTS_LANGUAGE"
-        android:summary="@string/preferences_general_tts_language_description"
-        android:title="@string/preferences_general_tts_language_title" />
-
 </PreferenceScreen>
